requirements.txt
google-generativeai
pydantic
yaml
deepagents
